cli - bin - tag - fix walk()
Yielding path at the beginning of walk() is necessary to handle
files. However, is isn't sufficient: recursive walk needs to
yield the current directory at each level of recursion.

If we yield dirpath, we should not directly yield path to avoid
duplicates. Thus we test if path is a directory before yielding
it.